+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Added a quick settings to timeline header for easier access
 - Added option to mark posts as sensitive by default
+- Added quick filters for notifications
+
 
 ## [2.3.0] - 2021-03-01
 ### Fixed
@@ -29,6 +31,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Added reason field for registration when approval is required
 - Group staff members by role in the About page
+
 
 ## [2.2.3] - 2021-01-18
 ### Added